Fix deleting from db

+++ Brio.Docs.External/CloudBase/CommonConnectionStorage.cs
@@ -18,12 +18,12 @@ namespace Brio.Docs.External.CloudBase
 
         public async Task<bool> DeleteFiles(string projectId, IEnumerable<ItemExternalDto> itemExternalDtos)
         {
-            var projectFiles = ItemsSyncHelper.GetProjectItems(projectId, cloudManager);
+            var projectFiles = await ItemsSyncHelper.GetProjectItems(projectId, cloudManager);
             var deletionResult = true;
             foreach (var item in itemExternalDtos)
             {
-                //if (!(await projectFiles).Any(f => f.ExternalID.Equals(item.ExternalID)))
-                //    return false;
+                if (!projectFiles.Any(f => f.ExternalID.Equals(item.ExternalID)))
+                    return false;
 
                 if (!string.IsNullOrWhiteSpace(item?.ExternalID))
                     deletionResult = await cloudManager.DeleteFile(item.ExternalID) && deletionResult;
